+++ src/main/java/com/WalkiePaw/presentation/domain/report/MemberReportController.java
@@ -31,31 +31,31 @@ public class MemberReportController {
     }
 
     @GetMapping("/{id}")
-    public ResponseEntity<MemberReportGetResponse> getMemberReport(final @PathVariable("id") Integer memberReportId) {
+    public ResponseEntity<MemberReportGetResponse> getMemberReport(final @PathVariable("id") Long memberReportId) {
         MemberReportGetResponse response = memberReportService.findById(memberReportId);
         return ResponseEntity.ok().body(response);
     }
 
     @PostMapping
     ResponseEntity<Void> addMemberReport(final @Validated @RequestBody MemberReportAddRequest request) {
-        Integer memberReportId = memberReportService.save(request);
+        Long memberReportId = memberReportService.save(request);
         return ResponseEntity.created(URI.create(MEMBER_REPORT_URL + memberReportId)).build();
     }
 
     @PatchMapping("/{id}")
-    ResponseEntity<Void> updateMemberReport(final @PathVariable("id") Integer memberReportId, final @Validated @RequestBody MemberReportUpdateRequest request) {
+    ResponseEntity<Void> updateMemberReport(final @PathVariable("id") Long memberReportId, final @Validated @RequestBody MemberReportUpdateRequest request) {
         memberReportService.update(memberReportId, request);
         return ResponseEntity.noContent().build();
     }
 
     @PatchMapping("/{id}/ban")
-    ResponseEntity<Void> banMemberReport(@PathVariable("id") final Integer memberReportId) {
+    ResponseEntity<Void> banMemberReport(@PathVariable("id") final Long memberReportId) {
         memberReportService.ban(memberReportId);
         return ResponseEntity.noContent().build();
     }
 
     @PatchMapping("/{id}/ignore")
-    ResponseEntity<Void> ignoreMemberReport(@PathVariable("id") final Integer memberReportId) {
+    ResponseEntity<Void> ignoreMemberReport(@PathVariable("id") final Long memberReportId) {
         memberReportService.ignore(memberReportId);
         return ResponseEntity.noContent().build();
     }
